backend/requirements.txt
Flask
Flask-Cors
Flask-SocketIO
python-engineio
python-socketio
numpy
scipy
mne
python-dotenv
eventlet
dnspython
